third_party/WebKit/LayoutTests/media/track/track-css-user-settings-override-author-settings-expected.txt

Test that WebVTT objects are being styled correctly based on user settings that should override author settings.
EVENT(canplaythrough)

EXPECTED (getComputedStyle(cue).color == 'rgb(255, 0, 0)') OK
EXPECTED (getComputedStyle(cue).backgroundColor == 'rgb(0, 128, 0)') OK
EXPECTED (getComputedStyle(cue).textShadow == 'rgb(0, 255, 0) 3px 3px 0px') OK
EXPECTED (getComputedStyle(cue).fontSize == '20px') OK
EXPECTED (getComputedStyle(cue).fontFamily == 'arial') OK
EXPECTED (getComputedStyle(cue).fontStyle == 'normal') OK
EXPECTED (getComputedStyle(cue).fontVariant == 'normal') OK
RUN(video.currentTime = 0.3)

EXPECTED (getComputedStyle(cue).color == 'rgb(0, 255, 255)') OK
EXPECTED (getComputedStyle(cue).backgroundColor == 'rgb(0, 128, 0)') OK
EXPECTED (getComputedStyle(cue).textShadow == 'rgb(255, 0, 0) 2px 2px 0px') OK
EXPECTED (getComputedStyle(cue).fontSize == '14px') OK
EXPECTED (getComputedStyle(cue).fontFamily == 'fantasy') OK
EXPECTED (getComputedStyle(cue).fontStyle == 'italic') OK
EXPECTED (getComputedStyle(cue).fontVariant == 'small-caps') OK
END OF TEST
